restructuring of task step implementations
