+++ grizzly/tasks/__init__.py
@@ -248,14 +248,14 @@ class template:
 
 
 from .request import RequestTask, RequestTaskHandlers, RequestTaskResponse
-from .wait import WaitTask
+from .wait_explicit import ExplicitWaitTask
 from .log_message import LogMessageTask
 from .transformer import TransformerTask
 from .until import UntilRequestTask
 from .date import DateTask
 from .async_group import AsyncRequestGroupTask
 from .timer import TimerTask
-from .task_wait import TaskWaitTask
+from .wait_between import WaitBetweenTask
 from .conditional import ConditionalTask
 from .loop import LoopTask
 from .set_variable import SetVariableTask
@@ -267,13 +267,13 @@ __all__ = [
     'RequestTaskResponse',
     'RequestTask',
     'LogMessageTask',
-    'WaitTask',
+    'ExplicitWaitTask',
     'TransformerTask',
     'UntilRequestTask',
     'DateTask',
     'AsyncRequestGroupTask',
     'TimerTask',
-    'TaskWaitTask',
+    'WaitBetweenTask',
     'ConditionalTask',
     'LoopTask',
     'SetVariableTask',
