+++ tests/unit/test_grizzly/steps/scenario/test_response.py
@@ -7,7 +7,7 @@ from parse import compile
 from grizzly.context import GrizzlyContext
 from grizzly.types import RequestMethod, ResponseTarget
 from grizzly.types.behave import Table, Row
-from grizzly.tasks import RequestTask, WaitTask
+from grizzly.tasks import RequestTask, ExplicitWaitTask
 from grizzly.steps import *  # pylint: disable=unused-wildcard-import  # noqa: F403
 
 from grizzly_extras.transformer import TransformerContentType
@@ -360,7 +360,7 @@ def test_step_response_content_type(behave_fixture: BehaveFixture) -> None:
         step_response_content_type(behave, TransformerContentType.JSON)
     assert 'There are no requests in the scenario' in str(ae)
 
-    grizzly.scenario.tasks.add(WaitTask(time_expression='1.0'))
+    grizzly.scenario.tasks.add(ExplicitWaitTask(time_expression='1.0'))
 
     with pytest.raises(AssertionError) as ae:
         step_response_content_type(behave, TransformerContentType.JSON)
