+++ tests/unit/test_grizzly/steps/test__helpers.py
@@ -14,7 +14,7 @@ from grizzly.context import GrizzlyContext, GrizzlyContextScenario
 from grizzly.tasks.async_group import AsyncRequestGroupTask
 from grizzly.types import RequestMethod, ResponseTarget, ResponseAction
 from grizzly.types.behave import Table, Row
-from grizzly.tasks import RequestTask, WaitTask
+from grizzly.tasks import RequestTask, ExplicitWaitTask
 from grizzly.testdata.utils import templatingfilter
 from grizzly.steps._helpers import (
     add_validation_handler,
@@ -151,7 +151,7 @@ def test_add_request_task(grizzly_fixture: GrizzlyFixture, tmp_path_factory: Tem
 
         if not as_async:
             tasks.clear()
-            tasks.append(WaitTask(time_expression='1.0'))
+            tasks.append(ExplicitWaitTask(time_expression='1.0'))
 
             with pytest.raises(ValueError) as e:
                 add_request_task(behave, method=RequestMethod.PUT, source='template.j2.json')
@@ -345,7 +345,7 @@ def test_add_save_handler(behave_fixture: BehaveFixture, locust_fixture: LocustF
     assert user.context_variables.get('test-variable-payload', 'payload') is None
 
     # previous non RequestTask task
-    tasks.append(WaitTask(time_expression='1.0'))
+    tasks.append(ExplicitWaitTask(time_expression='1.0'))
 
     grizzly.state.variables['test'] = 'none'
     with pytest.raises(ValueError):
