+++ tests/unit/test_grizzly/tasks/test_wait_between.py
@@ -1,20 +1,20 @@
 import pytest
 
 from locust.exception import MissingWaitTimeError
-from grizzly.tasks import TaskWaitTask
+from grizzly.tasks import WaitBetweenTask
 
 from tests.fixtures import GrizzlyFixture
 
 
-class TestTaskWaitTask:
+class TestWaitBetweenTask:
     def test___init__(self) -> None:
-        task_factory = TaskWaitTask(1.0)
+        task_factory = WaitBetweenTask(1.0)
 
         assert task_factory.min_time == 1.0
         assert task_factory.max_time is None
         assert task_factory.__template_attributes__ == set()
 
-        task_factory = TaskWaitTask(2.0, 13.0)
+        task_factory = WaitBetweenTask(2.0, 13.0)
         assert task_factory.min_time == 2.0
         assert task_factory.max_time == 13.0
 
@@ -27,14 +27,14 @@ class TestTaskWaitTask:
         with pytest.raises(MissingWaitTimeError):
             parent.wait_time()
 
-        task = TaskWaitTask(1.0, 12.0)()
+        task = WaitBetweenTask(1.0, 12.0)()
 
         task(parent)
 
         wait_time = parent.wait_time()
         assert wait_time >= 1.0 and wait_time <= 12.0
 
-        task = TaskWaitTask(13.0)()
+        task = WaitBetweenTask(13.0)()
 
         task(parent)
