+++ tests/unit/test_grizzly/test_locust.py
@@ -36,7 +36,7 @@ from grizzly.types import RequestDirection, RequestMethod, RequestType
 from grizzly.types.locust import Environment
 from grizzly.types.behave import Context, Scenario
 from grizzly.context import GrizzlyContext
-from grizzly.tasks import LogMessageTask, RequestTask, WaitTask
+from grizzly.tasks import LogMessageTask, RequestTask, ExplicitWaitTask
 from grizzly.tasks.clients import MessageQueueClientTask
 from grizzly.users import RestApiUser, MessageQueueUser
 from grizzly.scenarios import IteratorScenario
@@ -181,7 +181,7 @@ def test_setup_locust_scenarios(behave_fixture: BehaveFixture, noop_zmq: NoopZmq
 
     task = RequestTask(RequestMethod.GET, 'test-1', '/api/v1/test/1')
     grizzly.scenario.tasks.add(task)
-    grizzly.scenario.tasks.add(WaitTask(time_expression='1.5'))
+    grizzly.scenario.tasks.add(ExplicitWaitTask(time_expression='1.5'))
     grizzly.scenario.tasks.add(LogMessageTask(message='test message'))
 
     # incorrect user type
@@ -626,7 +626,7 @@ def test_run_worker(behave_fixture: BehaveFixture, capsys: CaptureFixture, mocke
     grizzly.scenarios.create(behave_fixture.create_scenario('test-non-mq'))
     grizzly.scenario.user.class_name = 'RestApiUser'
     grizzly.scenario.context['host'] = 'https://test.example.org'
-    grizzly.scenario.tasks.add(WaitTask(time_expression='1.5'))
+    grizzly.scenario.tasks.add(ExplicitWaitTask(time_expression='1.5'))
     task = RequestTask(RequestMethod.GET, 'test-1', '/api/v1/test/1')
     grizzly.scenario.tasks.add(task)
 
@@ -715,7 +715,7 @@ def test_run_master(behave_fixture: BehaveFixture, capsys: CaptureFixture, mocke
     grizzly.scenarios.create(behave_fixture.create_scenario('test'))
     grizzly.scenario.user.class_name = 'RestApiUser'
     grizzly.scenario.context['host'] = 'https://test.example.org'
-    grizzly.scenario.tasks.add(WaitTask(time_expression='1.5'))
+    grizzly.scenario.tasks.add(ExplicitWaitTask(time_expression='1.5'))
     task = RequestTask(RequestMethod.GET, 'test-1', '/api/v1/test/1')
     grizzly.scenario.tasks.add(task)
     grizzly.setup.spawn_rate = 1
